Abstract the Endpoint and Handler classes to a central utilities file
